@@ -1,67 +1,64 @@
-source 'https://rubygems.org'
+source "https://rubygems.org"
 
 def on_heroku?
-  ENV['DYNO']
+  ENV["DYNO"]
 end
 
-# Don't break multiple ruby versions in travis just because heroku requires a specific one
-ruby "1.9.3" if on_heroku?
+ruby "2.2.3"
 
-gem 'rails', '4.0.2'
+gem "rails", "4.2.5.1"
 
 # Bundle edge Rails instead:
-# gem 'rails', :git => 'git://github.com/rails/rails.git'
+# gem "rails", :git => "git://github.com/rails/rails.git"
 
-gem 'pg'
-gem 'slim'
-gem 'attr_extras'
-gem 'bootstrap_forms', github: 'barsoom/bootstrap_forms'
-gem 'faye-websocket'
-gem 'redis'
-gem 'thin'
+gem "pg"
+gem "slim"
+gem "attr_extras"
+gem "bootstrap_forms", github: "barsoom/bootstrap_forms"
+gem "faye-websocket"
+gem "redis"
+gem "thin"
+gem "httparty"
 
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails', '~> 4.0.0'
-  gem 'coffee-rails', '~> 4.0.0'
-  gem 'bootstrap-sass', '2.1.0.0'
+  gem "sass-rails", "~> 4.0.0"
+  gem "coffee-rails", "~> 4.0.0"
+  gem "bootstrap-sass", "2.1.0.0"
 
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
-  # gem 'therubyracer', :platforms => :ruby
+  # gem "therubyracer", :platforms => :ruby
 
-  gem 'uglifier', '>= 1.3.0'
+  gem "uglifier", ">= 1.3.0"
 end
 
 group :development, :test do
-  gem 'rspec-rails'
-
-  # Used by the report_build script.
-  gem 'httparty'
+  gem "rspec-rails"
 end
 
 group :test do
-  gem 'capybara'
-  gem 'factory_girl_rails'
+  gem "capybara"
+  gem "factory_girl_rails"
 end
 
 group :production do
-  gem 'rails_12factor'
+  gem "rails_12factor"
 end
 
-gem 'jquery-rails'
+gem "jquery-rails"
 
 # To use ActiveModel has_secure_password
-# gem 'bcrypt-ruby', '~> 3.0.0'
+# gem "bcrypt-ruby", "~> 3.0.0"
 
 # To use Jbuilder templates for JSON
-# gem 'jbuilder'
+# gem "jbuilder"
 
 # Use unicorn as the app server
-# gem 'unicorn'
+# gem "unicorn"
 
 # Deploy with Capistrano
-# gem 'capistrano'
+# gem "capistrano"
 
 # To use debugger
-# gem 'debugger'
+# gem "debugger"
