@@ -1,14 +1,12 @@
 require "spec_helper"
 
 describe UpdateBuildStatus do
-  let(:update_build_status) { described_class }
-
   context "when there are no previous builds" do
     it "adds a build, a revision and a project" do
       update_with name: "pipeline_tests", repository: "git@example.com:user/bar.git"
 
       builds = Build.all
-      builds.size.should == 1
+      expect(builds.size).to eq(1)
       build = builds.first
       expect(build.name).to eq("pipeline_tests")
       expect(build.status_url).to eq("http://example.com/builds/1")
@@ -22,9 +20,9 @@ describe UpdateBuildStatus do
       update_with status: "successful", status_url: "http://example.com/updated"
 
       builds = Build.all
-      builds.size.should == 1
-      builds.first.status.should == "successful"
-      builds.first.status_url.should == "http://example.com/updated"
+      expect(builds.size).to eq(1)
+      expect(builds.first.status).to eq("successful")
+      expect(builds.first.status_url).to eq("http://example.com/updated")
     end
   end
 
@@ -38,17 +36,24 @@ describe UpdateBuildStatus do
   end
 
   context "when there are more than App.builds_to_keep builds" do
-    it "removes the oldest build" do
-      App.stub(revisions_to_keep: 2)
-      update_with revision: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-      update_with revision: "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"
-      update_with revision: "cccccccccccccccccccccccccccccccccccccccc"
+    it "removes the builds that should not be kept" do
+      allow(App).to receive(:revisions_to_keep).and_return(4)
+      update_with revision: "0000000000000000000000000000000000000000"
+      update_with revision: "1111111111111111111111111111111111111111"
+      update_with revision: "2222222222222222222222222222222222222222"
+      update_with revision: "3333333333333333333333333333333333333333"
+
+      # Removes old revisions if the number is changed between updates
+      allow(App).to receive(:revisions_to_keep).and_return(2)
+      update_with revision: "4444444444444444444444444444444444444444"
 
       project = Project.last
       list = project.revisions.map(&:name).to_s
-      list.should_not include("a")
-      list.should include("b")
-      list.should include("c")
+      expect(list).not_to include("0")
+      expect(list).not_to include("1")
+      expect(list).not_to include("2")
+      expect(list).to include("3")
+      expect(list).to include("4")
 
       expect(Build.count).to eq(2)
     end
@@ -63,7 +68,7 @@ describe UpdateBuildStatus do
       status: "building",
     }.merge(custom)
 
-    update_build_status.run(
+    UpdateBuildStatus.call(
       attributes[:name],
       attributes[:repository],
       attributes[:revision],
