@@ -1,8 +1,6 @@
 class Api::BuildStatusesController < ApiController
-  before_filter :check_token
-
   def create
-    project = UpdateBuildStatus.run(
+    project = UpdateBuildStatus.call(
       params[:name],
       params[:repository],
       params[:revision],
@@ -12,13 +10,10 @@ class Api::BuildStatusesController < ApiController
 
     PushBackend.push({ project_id: project.id,
       html: render_to_string(partial: "projects/project",
-                             locals: { project: project }) })
+                             locals: { project: project, revision_amount: 2 }) })
+
+    PostStatusToWebhook.call(project)
+
     render nothing: true
-  end
-
-  private
-
-  def check_token
-    render nothing: true, status: :unauthorized unless App.api_token == params[:token]
   end
 end
