@@ -1,6 +1,17 @@
 class ProjectsController < WebController
+  before_action :get_projects
+
   def index
-    @projects = Project.all_sorted
+    revision_amount = 2
+    locals revision_amount: revision_amount
+  end
+
+  def show
+    project = Project.find(params[:id])
+    revision_amount = 15
+    locals :show,
+      {revision_amount: revision_amount,
+      project: project}
   end
 
   def edit
@@ -11,6 +22,7 @@ class ProjectsController < WebController
     project = Project.find(params[:id])
 
     if project.update_attributes(project_params)
+      PostStatusToWebhook.call(project)
       redirect_to root_path, notice: "Project updated."
     else
       @project = project
@@ -31,5 +43,9 @@ class ProjectsController < WebController
 
   def setup_menu
     active_menu_item_name :projects
+  end
+
+  def get_projects
+    @projects = Project.all_sorted
   end
 end
