@@ -1,13 +1,23 @@
+require "active_support/core_ext/module/delegation"
+require "merge_builds"
+
 class BuildPresenter
-  pattr_initialize :source_builds, :build_mappings
+  pattr_initialize :revision
 
   def list
-    builds = add_pending(source_builds)
+    builds = revision.builds
+    builds = add_pending(builds)
+    builds = timeout_builds(builds)
+    builds = change_status_for_fixed_builds(builds)
     builds = sort_by_mappings(builds)
-    map_names(builds)
+    builds = apply_mappings(builds)
+    merge_builds_with_the_same_name(builds)
   end
 
   private
+
+  delegate :build_mappings,
+    to: :revision
 
   def add_pending(builds)
     pending_mappings = build_mappings.select { |mapping|
@@ -15,10 +25,30 @@ class BuildPresenter
     }
 
     pending_builds = pending_mappings.map { |mapping|
-      new_build(mapping.from, "pending")
+      new_build(mapping.from, BuildStatus::PENDING)
     }
 
     builds + pending_builds
+  end
+
+  def timeout_builds(builds)
+    builds.map { |build|
+      if build.updated_at && build.updated_at < build_timeout.ago && build.status == BuildStatus::BUILDING
+        new_build(build.name, BuildStatus::PENDING, build)
+      else
+        build
+      end
+    }
+  end
+
+  def change_status_for_fixed_builds(builds)
+    builds.map { |build|
+      if newer_revision_fixes?(build)
+        new_build(build.name, BuildStatus::FIXED, build)
+      else
+        build
+      end
+    }
   end
 
   def sort_by_mappings(builds)
@@ -29,12 +59,17 @@ class BuildPresenter
     (mapped_builds + builds).uniq
   end
 
-  def map_names(builds)
+  def apply_mappings(builds)
     builds.map { |build|
-      mapping = mapping_for_build(build)
-      name = mapping ? mapping.to : build.name
+      name = mapping_for_build(build).to
       status = build.status
       new_build(name, status, build)
+    }
+  end
+
+  def merge_builds_with_the_same_name(builds)
+    builds.group_by(&:name).flat_map { |_, builds|
+      MergeBuilds.call(builds)
     }
   end
 
@@ -47,7 +82,20 @@ class BuildPresenter
   end
 
   def mapping_for_build(build)
-    build_mappings.find { |m| m.from == build.name }
+    build_mappings.find { |m| m.from == build.name } ||
+      BuildMapping.new(build.name, build.name)
+  end
+
+  def newer_revision_fixes?(build)
+    return false unless build.status == BuildStatus::FAILED
+
+    newer_revisions = revision.newer_revisions
+
+    newer_revisions.any? &&
+      newer_revisions.any? { |r|
+        newer_build = r.builds.find { |b| b.name == build.name }
+        newer_build && newer_build.status == BuildStatus::SUCCESSFUL
+      }
   end
 
   def new_build(name, status, old_build = nil)
@@ -55,5 +103,9 @@ class BuildPresenter
     build.name = name
     build.status = status
     build
+  end
+
+  def build_timeout
+    ENV.fetch("BUILD_TIMEOUT_IN_MINUTES", "60").to_i.minutes
   end
 end
